Added alpha and white point estimates to LischinskiTMO
Moved shared code between LischinskiTMO and ReinhardTMO to
input_estimates.hpp
